Crate lea_core

The core crate provides types and methods used by both the compiler and the VM.

Modules

fndata
Provides the FnData struct which is created by the compiler when emitting byte code and passed to the VM for execution.

limits
Implementation-dependent limitations on Lea programs

opcode
This module defines the Opcode type, which contains all opcodes handled by the VM.

Enums

Const
A constant as encountered in Lea source code

Number
Either an integer or a float. Converted as necessary.