Enum lea_core::opcode::Opcode

pub enum Opcode {
    MOV(u8, u8),
    LOADK(u8, u16),
    LOADNIL(u8, u8),
    LOADBOOL(u8, u8, bool),
    TABLE(u8),
    ARRAY(u8),
    FUNC(u8, u16),
    CALL(u8, u8, u8),
    RETURN(u8, u8),
    CLOSE(u8),
    VARARGS(u8, u8),
    JMP(i16),
    IF(u8, i16),
    IFNOT(u8, i16),
    FORCHECK(u8, u16),
    GETUPVAL(u8, u8),
    SETUPVAL(u8, u8),
    GETIDX(u8, u8, u8),
    SETIDX(u8, u8, u8),
    ADD(u8, u8, u8),
    SUB(u8, u8, u8),
    MUL(u8, u8, u8),
    DIV(u8, u8, u8),
    MOD(u8, u8, u8),
    POW(u8, u8, u8),
    EQ(u8, u8, u8),
    NEQ(u8, u8, u8),
    LEQ(u8, u8, u8),
    GEQ(u8, u8, u8),
    LESS(u8, u8, u8),
    GREATER(u8, u8, u8),
    BAND(u8, u8, u8),
    BOR(u8, u8, u8),
    BXOR(u8, u8, u8),
    SHIFTL(u8, u8, u8),
    SHIFTR(u8, u8, u8),
    CONCAT(u8, u8, u8),
    NEG(u8, u8),
    INV(u8, u8),
    LEN(u8, u8),
    NOT(u8, u8),
    INVALID,
    DEBUG(u8),
}

An opcode that can be executed by the VM. This enum is always 32 bits large.

Documentation syntax:

R[n] is the value stored in register (stack slot) n

U[n] is the current value of the upvalue with the index n

C[n] is the constant at index n (looked up in the current function's constant table)

PROTO[n] is function prototype number n

A, B, C are the values of the first, second and third u8 parameters (resp.)

Xu is the value of the operand of type u16 (eXtended unsigned)

Xs is the value of the operand of type i16 (eXtended signed)

PC is the program counter, which holds the number of the next opcode by default

Variants
MOV(u8, u8)

R[A] := R[B]

LOADK(u8, u16)

R[A] := C[Xu]

Loads the constant with index Xu into R[A].

LOADNIL(u8, u8)

R[A] := nil, R[A+1] := nil, ..., R[A+B] := nil

Assigns nil to the B+1 registers starting at A. If B is 0, assigns nil only to R[A].

LOADBOOL(u8, u8, bool)

R[A], R[A+1], ..., R[A+B] := C

Assigns a boolean to B+1 registers from R[A] to R[A+B]. If C is false, assign false to the registers, otherwise assign true.

TABLE(u8)

R[A] := {}

Creates a new table and assigns it to R[A].

ARRAY(u8)

R[A] := []

Creates a new array and assigns it to R[A].

FUNC(u8, u16)

R[A] := closure of PROTO[Xu]

Instantiates function prototype number Xu and stores a reference to the created closure in R[A]

CALL(u8, u8, u8)

R[A], R[A+1], ..., R[A+C-2] := R\[A](R[A+1], ..., R[A+B-1])

Calls R[A] with B-1 arguments stored in R[A+1] through R[A+B-1]. If B is 1, no arguments are passed. If B is 0, passes all arguments from A+1 to the top of the stack (this is used if the last argument is a function call, which may return an unknown number of values).

Stores C-1 return values in R[A] through R[A+C-2]. If C is 1, no return values are stored. If C is 0, stores all return values in R[A], R[A+1], ... (if necessary, expanding the stack to fit all return values in).

If the callee returns less than C-1 results, the rest is set to nil, except when C is 0, in which case 0 results may be returned (the top of the stack is set to A in that case).

This might invoke a metamethod if R[A] has a metatable with a __call field.

RETURN(u8, u8)

return R[A], R[A+1], ..., R[A+B-2]

Returns B-1 values stored in R[A] through R[A+B-2] to the caller. If B is 1, returns no values and just leaves the current function, returning control to the caller. If B is 0, returns all values from A to the top of the stack.

If the program's main function executes this opcode, the VM will pass the return values to the code that started the VM.

CLOSE(u8)

close_upval(R[A], R[A+1], ...)

Closes all open upvalues in registers A and higher.

VARARGS(u8, u8)

R[A], R[A+1], ..., R[A+B-1] := varargs

Stores B arguments passed as variable arguments (in ...) in R[A] through R[A+B-1]. If B is 0, stores all varargs in R[A], R[A+1], ... (expanding the stack if necessary).

JMP(i16)

PC += Xs

Unconditional jump. Offsets PC by Xs, which may be negative.

All jumps share the same property: If Xs is 0, PC will not be changed (the jump acts like a no-op). If Xs is -1, an endless loop is created.

IF(u8, i16)

if R[A]: PC += Xs

Conditional jump. If the value in R[A] is "truthy" (not false or nil), offsets PC by Xs.

IFNOT(u8, i16)

if not R[A]: PC += Xs

If the value in R[A] is false or nil, offsets PC by Xs.

FORCHECK(u8, u16)

assert: R[A], R[A+1], R[A+2] are numeric

if (R[A+1] >= 0 && R[A] > R[A+2]) || (R[A+1] < 0 && R[A] < R[A+2]): PC += Xu

Jumps out of a for-loop if the loop variable in R[A] is outside the range of the loop: If the loop step in R[A+1] is non-negative, this occurs exactly when the value is larger than the loop limit in R[A+2]. If the step is negative, this occurs when the value is smaller than the limit, since the loop runs backwards.

GETUPVAL(u8, u8)

R[A] := U[B]

SETUPVAL(u8, u8)

U[A] := R[B]

GETIDX(u8, u8, u8)

R[A] := R[B][R[C]]

Indexes R[B] with the value in R[C] and stores the result in R[A].

SETIDX(u8, u8, u8)

R[A][R[B]] := R[C]

Stores R[C] in R[A] at index R[B].

ADD(u8, u8, u8)

R[A] := R[B] + R[C]

All operator opcodes can call a metamethod, depending on the values in R[B] and R[C] and the specific operation.

SUB(u8, u8, u8)

R[A] := R[B] - R[C]

MUL(u8, u8, u8)

R[A] := R[B] * R[C]

DIV(u8, u8, u8)

R[A] := R[B] / R[C]

MOD(u8, u8, u8)

R[A] := R[B] % R[C]

POW(u8, u8, u8)

R[A] := R[B] ^ R[C]

EQ(u8, u8, u8)

R[A] := R[B] == R[C]

NEQ(u8, u8, u8)

R[A] := R[B] != R[C]

LEQ(u8, u8, u8)

R[A] := R[B] <= R[C]

GEQ(u8, u8, u8)

R[A] := R[B] >= R[C]

LESS(u8, u8, u8)

R[A] := R[B] < R[C]

GREATER(u8, u8, u8)

R[A] := R[B] > R[C]

BAND(u8, u8, u8)

R[A] := R[B] & R[C]

Bitwise and. Logical and / or (&& / ||) can shortcut and is implemented via test-and-jump. && / || can also not be overridden via a metamethod.

BOR(u8, u8, u8)

R[A] := R[B] | R[C]

BXOR(u8, u8, u8)

R[A] := R[B] ~ R[C]

SHIFTL(u8, u8, u8)

R[A] := R[B] << R[C]

SHIFTR(u8, u8, u8)

R[A] := R[B] >> R[C]

CONCAT(u8, u8, u8)

R[A] := R[B] .. R[C]

Concatenates the value in R[B] with the value in R[C]. Numbers are coerced to strings. If one of the values is neither a string nor a number, the __concat metamethod is called to convert the value.

NEG(u8, u8)

R[A] := -R[B]

INV(u8, u8)

R[A] := ~R[B]

LEN(u8, u8)

R[A] := #R[B]

NOT(u8, u8)

R[A] := !R[B]

The values false and nil will be converted to true, any other value will be converted to false. This does not call a metamethod.

INVALID

Causes a VM panic. Used by the emitter for opcodes that will be replaced later in the compilation (such as forward jumps).

DEBUG(u8)

Testing and debugging opcode that will print the Value stored in the given stack slot.